move processStructField to utils
